+++ jest.config.js
@@ -1,9 +1,15 @@
 /** @type {import('ts-jest').JestConfigWithTsJest} **/
 module.exports = {
     preset: 'ts-jest',
-    testEnvironment: 'node',
     rootDir: './test',
+    testEnvironment: 'node',
     transform: {
-        '^.+.tsx?$': ['ts-jest', {}],
+        '^.+\\.tsx?$': ['ts-jest', {}],
+    },
+    moduleNameMapper: {
+        '^(\\.{1,2}/.*)\\.js$': '$1',
+    },
+    testEnvironmentOptions: {
+        customExportConditions: ['node', 'node-addons'],
     },
 };