+++ src/YtdlCore.ts
@@ -1,9 +1,6 @@
-type YTDL_Constructor = Omit<YTDL_DownloadOptions, 'format'> & {
-    fetcher?: (url: URL | RequestInfo, options?: RequestInit) => Promise<Response>;
-    logDisplay?: Array<'debug' | 'info' | 'success' | 'warning' | 'error'>;
-};
+import UAParser from 'ua-parser-js';
 
-import type { YTDL_DownloadOptions, YTDL_GetInfoOptions, YTDL_VideoInfo, YTDL_OAuth2Credentials } from './types';
+import type { YTDL_Constructor, YTDL_DownloadOptions, YTDL_GetInfoOptions, YTDL_VideoInfo, YTDL_OAuth2Credentials, YTDL_DefaultStreamType, YTDL_NodejsStreamType } from './types';
 import type { InternalDownloadOptions } from './core/types';
 
 import { Platform } from './platforms/Platform';
@@ -12,21 +9,46 @@ import { download, downloadFromInfo } from './core/Download';
 import { getBasicInfo, getFullInfo } from './core/Info';
 import { getHtml5Player } from './core/Info/parser/Html5Player';
 import { OAuth2 } from './core/OAuth2';
+import { PlatformError } from './core/errors';
 
 import { Url } from './utils/Url';
 import { FormatUtils } from './utils/Format';
 import { VERSION } from './utils/Constants';
 import { Logger } from './utils/Log';
 
-const FileCache = Platform.getShim().fileCache;
+const SHIM = Platform.getShim(),
+    Cache = SHIM.cache,
+    FileCache = SHIM.fileCache;
 
 function isNodeVersionOk(version: string): boolean {
     return parseInt(version.replace('v', '').split('.')[0]) >= 16;
 }
 
-class YtdlCore {
-    public static writeStreamToFile?: (readableStream: ReadableStream, filePath: string) => Promise<void> = undefined;
+function isBrowserVersionOk(userAgent: string): boolean {
+    const PARSED = new UAParser(userAgent).getBrowser(),
+        VERSION = PARSED.version || '',
+        BROWSER_NAME = PARSED.name?.toLocaleLowerCase() || '';
 
+    if (BROWSER_NAME.includes('chrome')) {
+        return VERSION.startsWith('76');
+    } else if (BROWSER_NAME.includes('edge')) {
+        return VERSION.startsWith('80');
+    } else if (BROWSER_NAME.includes('firefox')) {
+        return VERSION.startsWith('78');
+    } else if (BROWSER_NAME.includes('safari')) {
+        return VERSION.startsWith('14');
+    } else if (BROWSER_NAME.includes('brave')) {
+        return VERSION.startsWith('1');
+    } else if (BROWSER_NAME.includes('opera')) {
+        return VERSION.startsWith('63');
+    } else {
+        Logger.debug('[ BrowserCheck ]: The specified UserAgent did not match the following browsers. (Chrome, Edge, FireFox, Safari, Brave, Opera)');
+    }
+
+    return true;
+}
+
+class YtdlCore {
     public static chooseFormat = FormatUtils.chooseFormat;
     public static filterFormats = FormatUtils.filterFormats;
 
@@ -65,9 +87,95 @@ class YtdlCore {
     public highWaterMark: YTDL_DownloadOptions['highWaterMark'] | undefined;
     public IPv6Block: YTDL_DownloadOptions['IPv6Block'] | undefined;
     public dlChunkSize: YTDL_DownloadOptions['dlChunkSize'] | undefined;
+    public streamType: YTDL_DownloadOptions['streamType'] = 'default';
 
     /* Metadata */
     public version = VERSION;
+
+    /* Constructor */
+    constructor({ hl, gl, rewriteRequest, poToken, disablePoTokenAutoGeneration, visitorData, includesPlayerAPIResponse, includesNextAPIResponse, includesOriginalFormatData, includesRelatedVideo, clients, disableDefaultClients, oauth2Credentials, parsesHLSFormat, originalProxy, quality, filter, excludingClients, includingClients, range, begin, liveBuffer, highWaterMark, IPv6Block, dlChunkSize, streamType, disableBasicCache, disableFileCache, fetcher, logDisplay, noUpdate }: YTDL_Constructor = {}) {
+        const SHIM = Platform.getShim();
+
+        /* Other Options */
+        const LOG_DISPLAY = (logDisplay === 'none' ? [] : logDisplay) || ['info', 'success', 'warning', 'error'];
+        SHIM.options.other.logDisplay = LOG_DISPLAY;
+        Logger.logDisplay = LOG_DISPLAY;
+
+        SHIM.options.other.noUpdate = noUpdate ?? false;
+
+        if (fetcher) {
+            SHIM.fetcher = fetcher;
+            SHIM.requestRelated.originalProxy = originalProxy;
+            SHIM.requestRelated.rewriteRequest = rewriteRequest;
+        }
+
+        if (disableBasicCache) {
+            Cache.disable();
+        }
+
+        if (disableFileCache) {
+            FileCache.disable();
+        }
+
+        /* Get Info Options */
+        this.hl = hl || 'en';
+        this.gl = gl || 'US';
+        this.rewriteRequest = rewriteRequest || undefined;
+        this.disablePoTokenAutoGeneration = disablePoTokenAutoGeneration ?? false;
+        this.includesPlayerAPIResponse = includesPlayerAPIResponse ?? false;
+        this.includesNextAPIResponse = includesNextAPIResponse ?? false;
+        this.includesOriginalFormatData = includesOriginalFormatData ?? false;
+        this.includesRelatedVideo = includesRelatedVideo ?? true;
+        this.clients = clients || undefined;
+        this.disableDefaultClients = disableDefaultClients ?? false;
+        this.parsesHLSFormat = parsesHLSFormat ?? false;
+
+        this.originalProxy = originalProxy || undefined;
+        if (this.originalProxy) {
+            Logger.debug(`<debug>"${this.originalProxy.base}"</debug> is used for <blue>API requests</blue>.`);
+            Logger.debug(`<debug>"${this.originalProxy.download}"</debug> is used for <blue>video downloads</blue>.`);
+            Logger.debug(`The query name <debug>"${this.originalProxy.urlQueryName || 'url'}"</debug> is used to specify the URL in the request. <blue>(?url=...)</blue>`);
+        }
+
+        this.setPoToken(poToken);
+        this.setVisitorData(visitorData);
+        this.setOAuth2(oauth2Credentials || null);
+
+        /* Format Selection Options */
+        this.quality = quality || undefined;
+        this.filter = filter || undefined;
+        this.excludingClients = excludingClients || [];
+        this.includingClients = includingClients || 'all';
+
+        /* Download Options */
+        this.range = range || undefined;
+        this.begin = begin || undefined;
+        this.liveBuffer = liveBuffer || undefined;
+        this.highWaterMark = highWaterMark || undefined;
+        this.IPv6Block = IPv6Block || undefined;
+        this.dlChunkSize = dlChunkSize || undefined;
+        this.streamType = streamType || 'default';
+
+        if (!this.disablePoTokenAutoGeneration) {
+            this.automaticallyGeneratePoToken();
+        }
+        this.initializeHtml5PlayerCache();
+
+        /* Version Check */
+        if (SHIM.runtime === 'default') {
+            if (!isNodeVersionOk(process.version)) {
+                throw new PlatformError(`You are using Node.js ${process.version} which is not supported. Minimum version required is v16.`);
+            }
+        } else if (SHIM.runtime === 'browser') {
+            if (!isBrowserVersionOk(navigator.userAgent)) {
+                throw new PlatformError(`The environment in which ytdl-core is currently running is not supported. ytdl-core has a minimum version that works with each browser due to limitations of the API it uses internally. Please see the README for details.`);
+            }
+        }
+
+        /* Load */
+        SHIM.options.download = { hl: this.hl, gl: this.gl, rewriteRequest: this.rewriteRequest, poToken: this.poToken, disablePoTokenAutoGeneration: this.disablePoTokenAutoGeneration, visitorData: this.visitorData, includesPlayerAPIResponse: this.includesPlayerAPIResponse, includesNextAPIResponse: this.includesNextAPIResponse, includesOriginalFormatData: this.includesOriginalFormatData, includesRelatedVideo: this.includesRelatedVideo, clients: this.clients, disableDefaultClients: this.disableDefaultClients, oauth2Credentials, parsesHLSFormat: this.parsesHLSFormat, originalProxy: this.originalProxy, quality: this.quality, filter: this.filter, excludingClients: this.excludingClients, includingClients: this.includingClients, range: this.range, begin: this.begin, liveBuffer: this.liveBuffer, highWaterMark: this.highWaterMark, IPv6Block: this.IPv6Block, dlChunkSize: this.dlChunkSize };
+        Platform.load(SHIM);
+    }
 
     /* Setup */
     private async setPoToken(poToken?: string) {
@@ -114,7 +222,7 @@ class YtdlCore {
 
     private automaticallyGeneratePoToken() {
         if (!this.poToken && !this.visitorData) {
-            Logger.debug('Since PoToken and VisitorData are not specified, they are generated automatically.');
+            Logger.debug('Since PoToken and VisitorData are <warning>not specified</warning>, they are generated <info>automatically</info>.');
 
             const generatePoToken = Platform.getShim().poToken;
 
@@ -136,69 +244,6 @@ class YtdlCore {
         if (!HTML5_PLAYER) {
             Logger.debug('To speed up processing, html5Player and signatureTimestamp are pre-fetched and cached.');
             getHtml5Player({});
-        }
-    }
-
-    constructor({ hl, gl, rewriteRequest, poToken, disablePoTokenAutoGeneration, visitorData, includesPlayerAPIResponse, includesNextAPIResponse, includesOriginalFormatData, includesRelatedVideo, clients, disableDefaultClients, oauth2Credentials, parsesHLSFormat, originalProxy, quality, filter, excludingClients, includingClients, range, begin, liveBuffer, highWaterMark, IPv6Block, dlChunkSize, disableFileCache, fetcher, logDisplay }: YTDL_Constructor = {}) {
-        /* Other Options */
-        Logger.logDisplay = logDisplay || ['info', 'success', 'warning', 'error'];
-        if (fetcher) {
-            const SHIM = Platform.getShim();
-            SHIM.fetcher = fetcher;
-            SHIM.requestRelated.originalProxy = originalProxy;
-            SHIM.requestRelated.rewriteRequest = rewriteRequest;
-            Platform.load(SHIM);
-        }
-        if (disableFileCache) {
-            FileCache.disable();
-        }
-
-        /* Get Info Options */
-        this.hl = hl || 'en';
-        this.gl = gl || 'US';
-        this.rewriteRequest = rewriteRequest || undefined;
-        this.disablePoTokenAutoGeneration = disablePoTokenAutoGeneration ?? false;
-        this.includesPlayerAPIResponse = includesPlayerAPIResponse ?? false;
-        this.includesNextAPIResponse = includesNextAPIResponse ?? false;
-        this.includesOriginalFormatData = includesOriginalFormatData ?? false;
-        this.includesRelatedVideo = includesRelatedVideo ?? true;
-        this.clients = clients || undefined;
-        this.disableDefaultClients = disableDefaultClients ?? false;
-        this.parsesHLSFormat = parsesHLSFormat ?? false;
-
-        this.originalProxy = originalProxy || undefined;
-        if (this.originalProxy) {
-            Logger.debug(`<debug>"${this.originalProxy.base}"</debug> is used for <blue>API requests</blue>.`);
-            Logger.debug(`<debug>"${this.originalProxy.download}"</debug> is used for <blue>video downloads</blue>.`);
-            Logger.debug(`The query name <debug>"${this.originalProxy.urlQueryName || 'url'}"</debug> is used to specify the URL in the request. <blue>(?url=...)</blue>`);
-        }
-
-        this.setPoToken(poToken);
-        this.setVisitorData(visitorData);
-        this.setOAuth2(oauth2Credentials || null);
-
-        /* Format Selection Options */
-        this.quality = quality || undefined;
-        this.filter = filter || undefined;
-        this.excludingClients = excludingClients || [];
-        this.includingClients = includingClients || 'all';
-
-        /* Download Options */
-        this.range = range || undefined;
-        this.begin = begin || undefined;
-        this.liveBuffer = liveBuffer || undefined;
-        this.highWaterMark = highWaterMark || undefined;
-        this.IPv6Block = IPv6Block || undefined;
-        this.dlChunkSize = dlChunkSize || undefined;
-
-        if (!this.disablePoTokenAutoGeneration) {
-            this.automaticallyGeneratePoToken();
-        }
-        this.initializeHtml5PlayerCache();
-
-        /* Version Check */
-        if (!isNodeVersionOk(process.version)) {
-            throw new Error(`You are using Node.js ${process.version} which is not supported. Minimum version required is v16.`);
         }
     }
 
@@ -234,6 +279,7 @@ class YtdlCore {
         INTERNAL_OPTIONS.highWaterMark = options.highWaterMark || this.highWaterMark || undefined;
         INTERNAL_OPTIONS.IPv6Block = options.IPv6Block || this.IPv6Block || undefined;
         INTERNAL_OPTIONS.dlChunkSize = options.dlChunkSize || this.dlChunkSize || undefined;
+        INTERNAL_OPTIONS.streamType = options.streamType || this.streamType || 'default';
 
         if (!INTERNAL_OPTIONS.oauth2 && options.oauth2Credentials) {
             Logger.warning('The OAuth2 token should be specified when instantiating the YtdlCore class, not as a function argument.');
@@ -245,12 +291,12 @@ class YtdlCore {
     }
 
     /** TIP: The options specified in new YtdlCore() are applied by default. (The function arguments specified will take precedence.) */
-    public download(link: string, options: YTDL_DownloadOptions = {}) {
-        return download(link, this.initializeOptions(options));
+    public download<T = YTDL_DefaultStreamType | YTDL_NodejsStreamType>(link: string, options: YTDL_DownloadOptions = {}): Promise<T extends YTDL_NodejsStreamType ? YTDL_NodejsStreamType : YTDL_DefaultStreamType> {
+        return download(link, this.initializeOptions(options)) as any;
     }
     /** TIP: The options specified in new YtdlCore() are applied by default. (The function arguments specified will take precedence.) */
-    public downloadFromInfo(info: YTDL_VideoInfo, options: YTDL_DownloadOptions = {}) {
-        return downloadFromInfo(info, this.initializeOptions(options));
+    public downloadFromInfo<T = YTDL_DefaultStreamType | YTDL_NodejsStreamType>(info: YTDL_VideoInfo, options: YTDL_DownloadOptions = {}): Promise<T extends YTDL_NodejsStreamType ? YTDL_NodejsStreamType : YTDL_DefaultStreamType> {
+        return downloadFromInfo(info, this.initializeOptions(options)) as any;
     }
 
     /** TIP: The options specified in new YtdlCore() are applied by default. (The function arguments specified will take precedence.) */