+++ src/utils/Constants.ts
@@ -1,5 +1,6 @@
 import pkg from '../../package.json';
 
+export const CURRENT_PLAYER_ID = '96d06116';
 export const VERSION = pkg.version;
 export const REPO_URL = 'https://github.com/ybd-project/ytdl-core';
 export const ISSUES_URL = 'https://github.com/ybd-project/ytdl-core/issues';