+++ v5/src/types/Options.ts
@@ -59,6 +59,9 @@ export type YTDL_GetInfoOptions = {
     /** You can specify a valid PoToken to avoid bot errors. */
     poToken?: string;
 
+    /** You can disable the automatic generation of PoToken. */
+    disablePoTokenAutoGeneration?: boolean;
+
     /** You can specify a valid VisitorData to avoid bot errors, etc. */
     visitorData?: string;